% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_function.R
\name{cat_function}
\alias{cat_function}
\title{Cat function}
\usage{
cat_function(love = TRUE)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats. Not sure how to do this.
}
\examples{
load(port25sb)
cat_function()
}
\keyword{cats}

