% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ffFactors}
\alias{ffFactors}
\title{Returns of Fama-French factors.}
\format{A data frame with 1068 rows and 5 variables:
\describe{
  \item{mktrf}{excess return of the market}
  \item{size}{size factor}
  \item{hml}{book-to-market factor}
  \item{umd}{momentum factor}
  \item{rf}{one-month risk-free rate}
}}
\source{
\url{http://mba.tuck.dartmouth.edu/pages/faculty/ken.french/data_library.html}
}
\usage{
ffFactors
}
\description{
A dataset containing monthly returns.
}
\keyword{datasets}

