% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{port25sb}
\alias{port25sb}
\title{Returns of 25 portfolios formed on size and book-to-market.}
\format{A data frame with 1068 rows and 25 variables:
\describe{
  \item{s1b1}{return of the small / low book-to-market portfolio}
  ...
  \item{s5b5}{return of the large / high book-to-market portfolio}
}}
\source{
\url{http://mba.tuck.dartmouth.edu/pages/faculty/ken.french/data_library.html}
}
\usage{
port25sb
}
\description{
A dataset containing monthly returns.
}
\keyword{datasets}

